feat(ui): validate-messages without side effects
