+++ .changeset/swift-bananas-know.md
@@ -1,0 +1,5 @@
+---
+'@lion/ui': patch
+---
+
+validate-messages without side effects
