+++ packages/ui/components/validate-messages/test/loadDefaultFeedbackMessages.test.js
@@ -1,6 +1,5 @@
 /* eslint-disable no-unused-vars, no-param-reassign */
 import { expect } from '@open-wc/testing';
-import { localize } from '@lion/ui/localize.js';
 import { Required } from '@lion/ui/form-core.js';
 import { loadDefaultFeedbackMessages } from '@lion/ui/validate-messages.js';
 
@@ -20,7 +19,7 @@ function getProtectedMembers(validatorEl) {
 }
 
 describe('loadDefaultFeedbackMessages', () => {
-  it('will set default feedback message for Required', async () => {
+  it('will work without providing a LocalizeManager instance', async () => {
     const el = new Required();
     const { getMessage } = getProtectedMembers(el);
     expect(await getMessage()).to.equals(
@@ -29,21 +28,5 @@ describe('loadDefaultFeedbackMessages', () => {
 
     loadDefaultFeedbackMessages();
     expect(await getMessage({ fieldName: 'password' })).to.equal('Please enter a(n) password.');
-  });
-
-  it('will await loading of translations when switching locale', async () => {
-    const el = new Required();
-    const { getMessage } = getProtectedMembers(el);
-    loadDefaultFeedbackMessages();
-    expect(await getMessage({ fieldName: 'password' })).to.equal('Please enter a(n) password.');
-    expect(await getMessage({ fieldName: 'user name' })).to.equal('Please enter a(n) user name.');
-
-    localize.locale = 'de-DE';
-    expect(await getMessage({ fieldName: 'Password' })).to.equal(
-      'Password muss ausgefüllt werden.',
-    );
-    expect(await getMessage({ fieldName: 'Benutzername' })).to.equal(
-      'Benutzername muss ausgefüllt werden.',
-    );
   });
 });
