+++ packages/ui/exports/validate-messages-no-side-effects.js
@@ -1,0 +1,1 @@
+export { loadDefaultFeedbackMessagesNoSideEffects } from '../components/validate-messages/src/loadDefaultFeedbackMessagesNoSideEffects.js';
